Started work on SamplingSettings/Noise Layers
