+++ animatediff/model_injection.py
@@ -11,9 +11,10 @@ from comfy.model_patcher import ModelPatcher
 
 from .motion_module_ad import AnimateDiffModel, has_mid_block, normalize_ad_state_dict
 from .logger import logger
-from .motion_utils import MotionCompatibilityError, NoiseType, normalize_min_max
+from .motion_utils import MotionCompatibilityError, normalize_min_max
 from .motion_lora import MotionLoraInfo, MotionLoraList
 from .model_utils import get_motion_lora_path, get_motion_model_path, get_sd_model_type
+from .sample_settings import SeedNoiseType
 
 
 # some motion_model casts here might fail if model becomes metatensor or is not castable;
@@ -33,7 +34,6 @@ class ModelPatcherAndInjector(ModelPatcher):
         # injection stuff
         self.motion_injection_params: InjectionParams = None
         self.motion_model: MotionModelPatcher = None
-        self.motion_model_sampling = None
     
     def model_patches_to(self, device):
         super().model_patches_to(device)
@@ -77,7 +77,6 @@ class ModelPatcherAndInjector(ModelPatcher):
         cloned = ModelPatcherAndInjector(self)
         cloned.motion_model = self.motion_model
         cloned.motion_injection_params = self.motion_injection_params
-        cloned.motion_model_sampling = self.motion_model_sampling
         return cloned
 
 
@@ -315,7 +314,8 @@ class InjectionParams:
         self.sync_context_to_pe = False
         self.loras: MotionLoraList = None
         self.motion_model_settings = MotionModelSettings()
-        self.noise_type: str = NoiseType.DEFAULT
+        self.sampling_settings = None
+        self.noise_type: str = SeedNoiseType.DEFAULT
         self.sub_idxs = None  # value should NOT be included in clone, so it will auto reset
