+++ animatediff/motion_utils.py
@@ -1,6 +1,3 @@
-from abc import ABC, abstractmethod
-from typing import Union
-
 import torch
 import torch.nn.functional as F
 from torch import Tensor, nn
@@ -9,10 +6,7 @@ import comfy.model_management as model_management
 import comfy.ops
 from comfy.cli_args import args
 from comfy.ldm.modules.attention import attention_basic, attention_pytorch, attention_split, attention_sub_quad, default
-from comfy.controlnet import broadcast_image_to
-from comfy.utils import repeat_to_batch_size
 
-from .motion_lora import MotionLoraInfo
 from .logger import logger
 
 
@@ -72,121 +66,10 @@ class CrossAttentionMM(nn.Module):
         return self.to_out(out)
 
 
-# super class to TemporalTransformer-like classes
-class TemporalTransformerGeneric:
-    def temporal_transformer_init(self, default_length: int):
-        self.video_length = default_length
-        self.full_length = default_length
-        self.scale_min = 1.0
-        self.scale_max = 1.0
-        self.raw_scale_mask: Union[Tensor, None] = None
-        self.temp_scale_mask: Union[Tensor, None] = None
-        self.sub_idxs: Union[list[int], None] = None
-        self.prev_hidden_states_batch = 0
-
-    def reset_temp_vars(self):
-        del self.temp_scale_mask
-        self.temp_scale_mask = None
-        self.prev_hidden_states_batch = 0
-
-    def get_scale_mask(self, hidden_states: Tensor) -> Union[Tensor, None]:
-        # if no raw mask, return None
-        if self.raw_scale_mask is None:
-            return None
-        shape = hidden_states.shape
-        batch, channel, height, width = shape
-        # if temp mask already calculated, return it
-        if self.temp_scale_mask != None:
-            # check if hidden_states batch matches
-            if batch == self.prev_hidden_states_batch:
-                if self.sub_idxs is not None:
-                    return self.temp_scale_mask[:, self.sub_idxs, :]
-                return self.temp_scale_mask
-            # if does not match, reset cached temp_scale_mask and recalculate it
-            del self.temp_scale_mask
-            self.temp_scale_mask = None
-        # otherwise, calculate temp mask
-        self.prev_hidden_states_batch = batch
-        mask = prepare_mask_batch(self.raw_scale_mask, shape=(self.full_length, 1, height, width))
-        mask = repeat_to_batch_size(mask, self.full_length)
-        # if mask not the same amount length as full length, make it match
-        if self.full_length != mask.shape[0]:
-            mask = broadcast_image_to(mask, self.full_length, 1)
-        # reshape mask to attention K shape (h*w, latent_count, 1)
-        batch, channel, height, width = mask.shape
-        # first, perform same operations as on hidden_states,
-        # turning (b, c, h, w) -> (b, h*w, c)
-        mask = mask.permute(0, 2, 3, 1).reshape(batch, height*width, channel)
-        # then, make it the same shape as attention's k, (h*w, b, c)
-        mask = mask.permute(1, 0, 2)
-        # make masks match the expected length of h*w
-        batched_number = shape[0] // self.video_length
-        if batched_number > 1:
-            mask = torch.cat([mask] * batched_number, dim=0)
-        # cache mask and set to proper device
-        self.temp_scale_mask = mask
-        # move temp_scale_mask to proper dtype + device
-        self.temp_scale_mask = self.temp_scale_mask.to(dtype=hidden_states.dtype, device=hidden_states.device)
-        # return subset of masks, if needed
-        if self.sub_idxs is not None:
-            return self.temp_scale_mask[:, self.sub_idxs, :]
-        return self.temp_scale_mask
-
-
 class BlockType:
     UP = "up"
     DOWN = "down"
     MID = "mid"
-
-
-class GenericMotionWrapper(nn.Module, ABC):
-    def __init__(self, mm_hash: str, mm_name: str, loras: list[MotionLoraInfo]):
-        super().__init__()
-        self.down_blocks: nn.ModuleList = None
-        self.up_blocks: nn.ModuleList = None
-        self.mid_block = None
-        self.mm_hash = mm_hash
-        self.mm_name = mm_name
-        self.version = "FILLTHISIN"
-        self.injector_version = "VERYIMPORTANT_FILLTHISIN"
-        self.AD_video_length: int = 0
-        self.loras = loras
-
-    def has_loras(self) -> bool:
-        # TODO: fix this to return False if has an empty list as well
-        # but only after implementing a fix for lowvram loading
-        return self.loras is not None
-
-    @abstractmethod
-    def set_video_length(self, video_length: int, full_length: int):
-        pass
-
-    @abstractmethod
-    def set_scale_multiplier(self, multiplier: Union[float, None]):
-        pass
-
-    @abstractmethod
-    def set_masks(self, masks: Tensor, min_val: float, max_val: float):
-        pass
-
-    @abstractmethod
-    def set_sub_idxs(self, sub_idxs: list[int]):
-        pass
-    
-    @abstractmethod
-    def reset_temp_vars(self):
-        pass
-
-    def reset_scale_multiplier(self):
-        self.set_scale_multiplier(None)
-
-    def reset_sub_idxs(self):
-        self.set_sub_idxs(None)
-
-    def reset(self):
-        self.reset_sub_idxs()
-        self.reset_scale_multiplier()
-        self.reset_temp_vars()
 
 
 class GroupNormAD(torch.nn.GroupNorm):
@@ -213,66 +96,6 @@ def prepare_mask_batch(mask: Tensor, shape: Tensor, multiplier: int=1, match_dim
     if match_dim1:
         mask = torch.cat([mask] * shape[1], dim=1)
     return mask
-
-
-class NoiseType:
-    DEFAULT = "default"
-    REPEATED = "repeated"
-    CONSTANT = "constant"
-    CONSTANT_ADDED = "constant_added"
-    AUTO1111 = "auto1111"
-
-    LIST = [DEFAULT, REPEATED, CONSTANT, CONSTANT_ADDED, AUTO1111]
-
-    @classmethod
-    def prepare_noise(cls, noise_type: str, latents: Tensor, noise: Tensor, context_length: int, seed: int):
-        if noise_type == cls.DEFAULT:
-            return noise
-        elif noise_type == cls.REPEATED:
-            return cls.prepare_noise_repeated(latents, noise, context_length, seed)
-        elif noise_type == cls.CONSTANT:
-            return cls.prepare_noise_constant(latents, noise, context_length, seed)
-        elif noise_type == cls.CONSTANT_ADDED:
-            new_noise = cls.prepare_noise_constant(latents, noise, context_length, seed)
-            noise_weight = 0.2
-            return noise * (1.0-(noise_weight/3)) + new_noise * noise_weight
-        elif noise_type == cls.AUTO1111:
-            return cls.prepare_noise_auto1111(latents, noise, context_length, seed)
-        logger.warning(f"Noise type {noise_type} not recognized, proceeding with default noise.")
-        return noise
-
-    @classmethod
-    def prepare_noise_repeated(cls, latents: Tensor, noise: Tensor, context_length: int, seed: int):
-        if not context_length:
-            return noise
-        length = latents.shape[0]
-        generator = torch.manual_seed(seed)
-        noise = torch.randn(latents.size(), dtype=latents.dtype, layout=latents.layout, generator=generator, device="cpu")
-        noise_set = noise[:context_length]
-        cat_count = (length // context_length) + 1
-        noise_set = torch.cat([noise_set] * cat_count, dim=0)
-        noise_set = noise_set[:length]
-        return noise_set
-
-    @classmethod
-    def prepare_noise_constant(cls, latents: Tensor, noise: Tensor, context_length: int, seed: int):
-        length = latents.shape[0]
-        single_shape = (1, latents.shape[1], latents.shape[2], latents.shape[3])
-        generator = torch.manual_seed(seed)
-        noise = torch.randn(single_shape, dtype=latents.dtype, layout=latents.layout, generator=generator, device="cpu")
-        return torch.cat([noise] * length, dim=0)
-
-    @classmethod
-    def prepare_noise_auto1111(cls, latents: Tensor, noise: Tensor, context_length: int, seed: int):
-        # auto1111 applies growing seeds for a batch
-        length = latents.shape[0]
-        single_shape = (1, latents.shape[1], latents.shape[2], latents.shape[3])
-        all_noises = []
-        # i starts at 0
-        for i in range(length):
-            generator = torch.manual_seed(seed+i)
-            all_noises.append(torch.randn(single_shape, dtype=latents.dtype, layout=latents.layout, generator=generator, device="cpu"))
-        return torch.cat(all_noises, dim=0)
 
 
 class MotionCompatibilityError(ValueError):
