+++ animatediff/nodes.py
@@ -7,9 +7,9 @@ from comfy.model_patcher import ModelPatcher
 from .context import ContextOptions, ContextSchedules, UniformContextOptions
 from .logger import logger
 from .model_utils import BetaSchedules, get_available_motion_loras, get_available_motion_models, get_motion_lora_path
-from .motion_utils import NoiseType
 from .motion_lora import MotionLoraInfo, MotionLoraList
 from .model_injection import InjectionParams, ModelPatcherAndInjector, MotionModelSettings, load_motion_module
+from .sample_settings import SeedNoiseType
 from .sampling import motion_sample_factory
 
 from .nodes_extras import AnimateDiffUnload, EmptyLatentImageLarge, CheckpointLoaderSimpleWithNoiseSelect
@@ -205,7 +205,7 @@ class AnimateDiffUniformContextOptionsExperimental:
                 "context_overlap": ("INT", {"default": 0, "min": 0, "max": 32}), # if new motion modules come out
                 "context_schedule": (ContextSchedules.CONTEXT_SCHEDULE_LIST,),
                 "closed_loop": ("BOOLEAN", {"default": False},),
-                "noise_override": (NoiseType.LIST,)
+                "noise_override": (SeedNoiseType.LIST,)
                 #"sync_context_to_pe": ("BOOLEAN", {"default": False},),
             },
         }
